+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 This is a high-level changelog for each released versions of the plugin.
 For a more detailed list of past and incoming changes, see the commit history.
 
+Main
+------
+- Added `create_channel`, `get_channel`, `update_channel` and `delete_channel` to `DiscordBot`
+- Added `get_channel_messages`, `get_channel_message` to `DiscordBot`
+
 1.2.2
 ------
 - Added `create_role`, `update_role` and `delete_role` to `DiscordBot`